+++ examples/producer.rs
@@ -1,61 +1,34 @@
 use bytes::Bytes;
+use futures::{stream::iter, StreamExt};
 use samsa::prelude::{ProduceMessage, ProducerBuilder};
 
 #[tokio::main]
 async fn main() -> Result<(), ()> {
     tracing_subscriber::fmt::init();
     let bootstrap_addrs = vec!["127.0.0.1:9092".to_string()];
-    let topics = vec!["purchases".to_string()];
+    let topic_name = "my-topic";
+
+    let stream = iter(vec![0].into_iter()).cycle().enumerate().map(|(i, _)| {
+        let partition_id = (i % 4) as i32;
+        ProduceMessage {
+            topic: topic_name.to_string(),
+            partition_id,
+            key: Some(Bytes::from_static(b"Tester")),
+            value: Some(Bytes::from_static(b"Value")),
+            headers: vec![],
+        }
+    });
 
     tracing::info!("Connecting to cluster");
-    let producer_client = ProducerBuilder::new(bootstrap_addrs, topics)
+    ProducerBuilder::new(bootstrap_addrs, vec![topic_name.to_string()])
         .await
         .map_err(|err| tracing::error!("{:?}", err))?
-        .build()
+        .build_from_stream(stream)
         .await;
 
-    let topic_name = "purchases";
-    let mut partition_id = 0;
+    // keep the original task alive
+    // TODO: fix this
+    tokio::time::sleep(tokio::time::Duration::MAX).await;
 
-    loop {
-        if partition_id <= 2 {
-            partition_id += 1
-        } else {
-            partition_id = 0
-        };
-
-        tracing::info!(
-            "Producing 3 records to topic {} partition {}",
-            topic_name.to_string(),
-            partition_id
-        );
-
-        producer_client
-            .produce(ProduceMessage {
-                topic: topic_name.to_string(),
-                partition_id,
-                key: Some(Bytes::from_static(b"Tester")),
-                value: Some(Bytes::from_static(b"Value 1")),
-                headers: vec![],
-            })
-            .await;
-        producer_client
-            .produce(ProduceMessage {
-                topic: topic_name.to_string(),
-                partition_id,
-                key: Some(Bytes::from_static(b"Tester")),
-                value: Some(Bytes::from_static(b"Value 2")),
-                headers: vec![],
-            })
-            .await;
-        producer_client
-            .produce(ProduceMessage {
-                topic: topic_name.to_string(),
-                partition_id,
-                key: Some(Bytes::from_static(b"Tester")),
-                value: Some(Bytes::from_static(b"Value 3")),
-                headers: vec![],
-            })
-            .await;
-    }
+    Ok(())
 }